Update no. 1
Rewrited ZO-AccSGD and ZO-VARAG methods, conducted several experiments. There is a strong discrepancy in our method after reaching the minimum value, it is unclear why this is happening. To solve this problem, I tried different batchsizes and increased/decreased the number of repeated oracle calls per iteration (direction variable). This has not changed the global model of behavior.